Avoid the weird shit markdown does to entities
The Python markdown package does some weird shit to HTML entities, which was resulting in strange characters being displayed in title attributes when HTML entities were expected. By moving the WikiLinks extension higher in the order of extensions, we take care of things before the markdown process has a change to munge the entities and everything works as expected.

...well, after we unescape the entities that Bleach was already escaping for us, of course!